use shared examples for tests
